+++ onmt/utils/Cuda.lua
@@ -117,6 +117,8 @@ function Cuda.convert(obj)
     if Cuda.activated and obj.cuda ~= nil then
       if objtype:find('torch%..*LongTensor') then
         return obj:type('torch.CudaLongTensor')
+      elseif objtype:find('torch%..*ByteTensor') then
+        return obj:type('torch.CudaByteTensor')
       elseif Cuda.fp16 then
         return obj:type('torch.CudaHalfTensor')
       else
@@ -126,6 +128,8 @@ function Cuda.convert(obj)
       -- Defaults to float instead of double.
       if objtype:find('torch%..*LongTensor') then
         return obj:type('torch.LongTensor')
+      elseif objtype:find('torch%..*ByteTensor') then
+        return obj:type('torch.ByteTensor')
       else
         return obj:type('torch.FloatTensor')
       end