+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Fixes and improvements
 
-* Fix beam-size 1 broken with lexical constraints
+* Misc fixes on lexical beam search
 * Fix batch size non function with `rest_translation_server.lua`
 * Introduce `-tokenizer max` option to scorer for evaluation on non tokenized test data.
 * Fix non deterministic inference of language models